+++ CHANGELOG.md
@@ -1,14 +1,29 @@
-## 1.58.2 (Aug 18, 2021)
+## 1.58.3 (Aug 18, 2021)
+
 ENHANCEMENTS:
+
+* Extend kubernetes node instance disk allow types
+* Resource `resource_tc_kubernetes_cluster_attachment.go` add `disk_partition` field
+
+BUGFIXES:
+
+* Resource `resource_tc_kubernetes_cluster` fix `auto_format_and_mount` default value to `false`
+
+## 1.58.2 (Aug 18, 2021)
+
+ENHANCEMENTS:
+
 * Resource `resource_tc_kubernetes_cluster.go` tke instance creation - add RuntimeVersion param
 * Resource `resource_tc_kubernetes_cluster.go` extend worker instance data disk mount settings
 
 ## 1.58.1 (Aug 17, 2021)
 
 ENHANCEMENTS:
+
 * Resource `resource_tc_mysql_instance.go` add cpu params for mysql
 
 BUG_FIXES:
+
 * Resource `resource_tc_instance.go` fix read cvm data_disks bug
 
 ## 1.58.0 (Aug 11, 2021)